Inline SVG in html document
Generate inline SVG without the need for inserting a 'raw' node.
It removes the inline base64 encoded SVG in a <img> tag as this did now allow style variations.
The css styling has been removed from the embedded SVG to enable site-wide control of the svgbob styling.
This allows for dealing with cases such as dark mode etc

Signed-off-by: Matt Spencer matthew@thespencers.me.uk